+++ src/main/java/me/pajic/tiered_backpacks/mixin/ArmorSlotMixin.java
@@ -2,13 +2,10 @@ package me.pajic.tiered_backpacks.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import com.llamalad7.mixinextras.sugar.Local;
-import me.pajic.tiered_backpacks.TieredBackpacks;
 import me.pajic.tiered_backpacks.util.BackpackUtil;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.ArmorSlot;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.component.ItemContainerContents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
@@ -20,7 +17,6 @@ public class ArmorSlotMixin {
             at = @At("RETURN")
     )
     private boolean modifyMayPickup(boolean original, @Local(name = "itemStack") ItemStack itemStack, @Local(argsOnly = true, name = "player") Player player) {
-        return !player.isCreative() && TieredBackpacks.CONFIG.preventUnequipWhenNotEmpty.get() && BackpackUtil.isValidContainerHolder(itemStack) ?
-				itemStack.getOrDefault(DataComponents.CONTAINER, ItemContainerContents.EMPTY) == ItemContainerContents.EMPTY : original;
+		return BackpackUtil.canUnequipBackpack(player, itemStack) && original;
     }
 }