+++ src/main/java/me/pajic/tiered_backpacks/util/TrinketsCompat.java
@@ -1,6 +1,8 @@
 package me.pajic.tiered_backpacks.util;
 
 import eu.pb4.trinkets.api.TrinketsApi;
+import me.pajic.tiered_backpacks.item.BackpackItem;
+import me.pajic.tiered_backpacks.item.BackpackTrinketItem;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 
@@ -10,5 +12,9 @@ public class TrinketsCompat {
 		var list = TrinketsApi.getAttachment(entity).getEquipped(stack -> stack.is(BackpackUtil.BACKPACKS));
 		if (!list.isEmpty()) return list.getFirst().getB();
 		return ItemStack.EMPTY;
+	}
+
+	public static BackpackItem makeTrinketBackpack(BackpackTier tier) {
+		return new BackpackTrinketItem(tier);
 	}
 }